w3c-reports/12092014/mac/runned-tests/webaudio/the-audio-api/the-delaynode-interface/idl-test-expected.txt

CONSOLE MESSAGE: line 188: callback not yet supported
CONSOLE MESSAGE: line 188: callback not yet supported
CONSOLE MESSAGE: Deprecated attempt to access property 'delayTime' on a non-DelayNode object.
interface DelayNode : AudioNode {

    readonly attribute AudioParam delayTime;

};

FAIL DelayNode interface: existence and properties of interface object assert_equals: prototype of window's property "DelayNode" is not Function.prototype expected (function) function "function () {
    [native code]
}" but got (object) object "[object Object]"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/resources/idlharness.js:991:22
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:947:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/webaudio/the-audio-api/the-delaynode-interface/idl-test.html:148:17
global code@http://localhost:8800/webaudio/the-audio-api/the-delaynode-interface/idl-test.html:149:3)
PASS DelayNode interface object length 
FAIL DelayNode interface: existence and properties of interface prototype object assert_false: DelayNode.prototype is enumerable expected false got true(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_false@http://localhost:8800/resources/testharness.js:306:15
http://localhost:8800/resources/idlharness.js:1094:21
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1071:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/webaudio/the-audio-api/the-delaynode-interface/idl-test.html:148:17
global code@http://localhost:8800/webaudio/the-audio-api/the-delaynode-interface/idl-test.html:149:3)
FAIL DelayNode interface: existence and properties of interface prototype object's "constructor" property assert_true: DelayNode.prototype.constructor is not writable expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1183:20
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1159:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/webaudio/the-audio-api/the-delaynode-interface/idl-test.html:148:17
global code@http://localhost:8800/webaudio/the-audio-api/the-delaynode-interface/idl-test.html:149:3)
FAIL DelayNode interface: attribute delayTime assert_throws: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1270:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
http://localhost:8800/webaudio/the-audio-api/the-delaynode-interface/idl-test.html:148:17
global code@http://localhost:8800/webaudio/the-audio-api/the-delaynode-interface/idl-test.html:149:3)
PASS DelayNode must be primary interface of delay_node 
PASS Stringification of delay_node 
PASS DelayNode interface: delay_node must inherit property "delayTime" with the proper type (0) 
PASS AudioNode interface: delay_node must inherit property "connect" with the proper type (0) 
PASS AudioNode interface: calling connect(AudioNode,unsigned long,unsigned long) on delay_node with too few arguments must throw TypeError 
PASS AudioNode interface: delay_node must inherit property "connect" with the proper type (1) 
PASS AudioNode interface: calling connect(AudioParam,unsigned long) on delay_node with too few arguments must throw TypeError 
PASS AudioNode interface: delay_node must inherit property "disconnect" with the proper type (2) 
PASS AudioNode interface: calling disconnect(unsigned long) on delay_node with too few arguments must throw TypeError 
PASS AudioNode interface: delay_node must inherit property "context" with the proper type (3) 
PASS AudioNode interface: delay_node must inherit property "numberOfInputs" with the proper type (4) 
PASS AudioNode interface: delay_node must inherit property "numberOfOutputs" with the proper type (5) 
PASS AudioNode interface: delay_node must inherit property "channelCount" with the proper type (6) 
PASS AudioNode interface: delay_node must inherit property "channelCountMode" with the proper type (7) 
PASS AudioNode interface: delay_node must inherit property "channelInterpretation" with the proper type (8) 
PASS EventTarget interface: delay_node must inherit property "addEventListener" with the proper type (0) 
FAIL EventTarget interface: calling addEventListener(DOMString,EventListener,boolean) on delay_node with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1640:34
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
http://localhost:8800/webaudio/the-audio-api/the-delaynode-interface/idl-test.html:148:17
global code@http://localhost:8800/webaudio/the-audio-api/the-delaynode-interface/idl-test.html:149:3)
PASS EventTarget interface: delay_node must inherit property "removeEventListener" with the proper type (1) 
FAIL EventTarget interface: calling removeEventListener(DOMString,EventListener,boolean) on delay_node with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1640:34
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
http://localhost:8800/webaudio/the-audio-api/the-delaynode-interface/idl-test.html:148:17
global code@http://localhost:8800/webaudio/the-audio-api/the-delaynode-interface/idl-test.html:149:3)
PASS EventTarget interface: delay_node must inherit property "dispatchEvent" with the proper type (2) 
PASS EventTarget interface: calling dispatchEvent(Event) on delay_node with too few arguments must throw TypeError 
